requirements.txt
crewai>=0.11.0
python-dotenv>=0.19.0
langchain-openai>=0.0.5

# Web Scraping and Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml>=5.3.0

# Social Media
twikit>=0.1.0
aiohttp>=3.8.0

# News and Content
newsapi-python>=0.2.7

# AI and Language Processing
openai>=1.3.0
litellm>=1.0.0

# Database
sqlalchemy>=2.0.0
pydantic>=2.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
